fix(multi): 多通道可视化适配与时间精度修复 [core:v1.5.1, web:v1.9.9]
core模块修复:
- Link_State_Visualizer重构为多网络类型+多通道架构，支持REQ/RSP/DATA网络的多通道切换
- 修复时间转换精度问题：int()改为round()，避免浮点除法精度丢失
- IPInterface移除单网络兼容引用(req_network等)，统一使用多网络列表(req_networks)
- StatsMixin适配多通道架构，同步channel buffer到所有网络实例
- 数据收集器、分析器和可视化器适配新的多通道数据结构

web平台优化:
- waveform API性能优化：预计算flit的start_time字段，避免重复JSON解析
- 使用itertuples替代iterrows提升DataFrame遍历性能
- 修复column_presets路径配置错误（多算了一层parent）
- WaveformResponse增加分页信息字段(total_packets, offset, returned_packets)

其他:
- .gitignore添加.env忽略规则
- 新增column_presets配置目录和多通道测试用例

+++ src/analysis/exporters.py
@@ -1650,7 +1650,7 @@ class JSONExporter:
 class ParquetExporter:
     """Parquet格式导出器 - 用于波形数据存储"""
 
-    def __init__(self, network_frequency: int = 2):
+    def __init__(self, network_frequency: float = 2.0):
         """
         初始化Parquet导出器
